requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
arrow=1.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backoff=2.2.1=pypi_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bezier=2023.7.28=pypi_0
blessed=1.20.0=pypi_0
blinker=1.7.0=pypi_0
boto3=1.33.11=pypi_0
botocore=1.33.11=pypi_0
ca-certificates=2023.08.22=hca03da5_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
croniter=1.4.1=pypi_0
cycler=0.12.1=pypi_0
dateutils=0.6.12=pypi_0
deepdiff=6.7.1=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.104.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.46.0=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.12.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
h11=0.14.0=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib-resources=6.1.1=pypi_0
inquirer=3.1.4=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.2=pypi_0
kiwisolver=1.4.5=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
lightning=2.1.2=pypi_0
lightning-app=2.1.2=pypi_0
lightning-cloud=0.5.52=pypi_0
lightning-utilities=0.10.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.4=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
opencv-python=4.8.1.78=pypi_0
openssl=3.0.12=h1a28f6b_0
ordered-set=4.1.0=pypi_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py38hca03da5_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.6=pypi_0
pyarrow=14.0.1=pypi_0
pydantic=2.5.2=pypi_0
pydantic-core=2.14.5=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pyjwt=2.8.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.18=hb885b13_0
python-dateutil=2.8.2=pypi_0
python-editor=1.0.4=pypi_0
python-multipart=0.0.6=pypi_0
pytorch-lightning=2.1.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readchar=4.0.5=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.32.0=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.2=pypi_0
s3transfer=0.8.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.0=pypi_0
segment-anything=1.0=pypi_0
setuptools=68.0.0=py38hca03da5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h80987f9_0
starlette=0.27.0=pypi_0
starsessions=1.3.0=pypi_0
streamlit=1.29.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tk=8.6.12=hb8d0fd4_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.1=pypi_0
torchaudio=2.1.1=pypi_0
torchmetrics=1.2.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
urllib3=1.26.18=pypi_0
uvicorn=0.24.0.post1=pypi_0
validators=0.22.0=pypi_0
wcwidth=0.2.12=pypi_0
websocket-client=1.7.0=pypi_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py38hca03da5_0
xz=5.4.5=h80987f9_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5a0b063_0
